move cluster checks handler installation to start server
